utils: Move I/O redirection related code to its own file
This continues the effort so modularize the monolithic "Utils.kt" file.
