+++ utils/src/test/kotlin/UtilsTest.kt
@@ -425,72 +425,4 @@ class UtilsTest : WordSpec({
             }
         }
     }
-
-    "redirecting output" should {
-        // Use a relatively large number of lines that results in more than 64k to be written to test against the pipe
-        // buffer limit on Linux, see https://unix.stackexchange.com/a/11954/53328.
-        val numberOfLines = 10000
-
-        "work for stdout only" {
-            val stdout = redirectStdout {
-                for (i in 1..numberOfLines) System.out.println("stdout: $i")
-            }
-
-            // The last printed line has a newline, resulting in a trailing blank line.
-            val stdoutLines = stdout.lines().dropLast(1)
-            stdoutLines.count() shouldBe numberOfLines
-            stdoutLines.last() shouldBe "stdout: $numberOfLines"
-        }
-
-        "work for stderr only" {
-            val stderr = redirectStderr {
-                for (i in 1..numberOfLines) System.err.println("stderr: $i")
-            }
-
-            // The last printed line has a newline, resulting in a trailing blank line.
-            val stderrLines = stderr.lines().dropLast(1)
-            stderrLines.count() shouldBe numberOfLines
-            stderrLines.last() shouldBe "stderr: $numberOfLines"
-        }
-
-        "work for stdout and stderr at the same time" {
-            var stderr = ""
-            val stdout = redirectStdout {
-                stderr = redirectStderr {
-                    for (i in 1..numberOfLines) {
-                        System.out.println("stdout: $i")
-                        System.err.println("stderr: $i")
-                    }
-                }
-            }
-
-            // The last printed line has a newline, resulting in a trailing blank line.
-            val stdoutLines = stdout.lines().dropLast(1)
-            stdoutLines.count() shouldBe numberOfLines
-            stdoutLines.last() shouldBe "stdout: $numberOfLines"
-
-            // The last printed line has a newline, resulting in a trailing blank line.
-            val stderrLines = stderr.lines().dropLast(1)
-            stderrLines.count() shouldBe numberOfLines
-            stderrLines.last() shouldBe "stderr: $numberOfLines"
-        }
-
-        "work when trapping exit calls" {
-            var exitCode : Int? = null
-
-            val stdout = redirectStdout {
-                exitCode = trapSystemExitCall {
-                    for (i in 1..numberOfLines) System.out.println("stdout: $i")
-                    System.exit(42)
-                }
-            }
-
-            exitCode shouldBe 42
-
-            // The last printed line has a newline, resulting in a trailing blank line.
-            val stdoutLines = stdout.lines().dropLast(1)
-            stdoutLines.count() shouldBe numberOfLines
-            stdoutLines.last() shouldBe "stdout: $numberOfLines"
-        }
-    }
 })
